+++ iris-mpc/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod aws;
+pub mod batch;
 pub mod client;
 pub mod iris_processing;
+pub mod message_receiver;
 pub mod utils;
