+++ iris-mpc/src/utils.rs
@@ -144,3 +144,15 @@ pub async fn initialize_chacha_seeds(config: Config) -> eyre::Result<([u32; 8], 
 
     Ok(chacha_seeds)
 }
+
+pub fn get_check_addresses(
+    hostnames: Vec<String>,
+    ports: Vec<String>,
+    endpoint: &str,
+) -> Vec<String> {
+    hostnames
+        .iter()
+        .zip(ports.iter())
+        .map(|(host, port)| format!("http://{}:{}/{}", host, port, endpoint))
+        .collect::<Vec<String>>()
+}
